refactor(api): extract L5RCMContext class to its own module
Phase 5 step 1 — preparation. Rename the L5RCMAPI class to
L5RCMContext and move it out of l5r/api/__init__.py into
l5r/api/context.py. Keep L5RCMAPI = L5RCMContext as a back-compat
alias for any external code referring to the old name.

No behavior change: __api is still a module-level instance and every
l5r.api.character.* / l5r.api.data.* function still reads it via the
same module-level reference. This commit just gives the class a
distinct, more accurate name and an isolated home, so subsequent
steps can grow it (add ctx threading, per-call instantiation) without
touching the rest of l5r/api/__init__.py.

The reload() method previously depended on get_user_data_path (defined
in l5r/api/__init__.py). To avoid a circular import, reload() now
takes that helper as a parameter; the __api singleton has a small
lambda wrapper bound in __init__.py so existing callers (__api.reload())
work unchanged.
